included functionality to the template, corrected class to ClassName in Wave.js, and commented in Home.js the Plane variable that was declared but never used
