+++ client/src/pages/Home.js
@@ -2,7 +2,7 @@ import React from "react";
 import Cart from "../components/Cart";
 import Header from "../components/Header/Header";
 import IconPaw from "../components/Header/IconPaw";
-import Plane from "../components/elements/Plane";
+//import Plane from "../components/elements/Plane";
 
 const Home = () => {
   return (
